fix: remove all deprecated members, update tests accordingly
