+++ flutter_secure_storage/android/src/main/java/com/it_nomads/fluttersecurestorage/FlutterSecureStorageConfig.java
@@ -17,7 +17,6 @@ public class FlutterSecureStorageConfig {
     private static final Boolean DEFAULT_DELETE_ON_FAILURE = false;
     private static final Boolean DEFAULT_MIGRATE_ON_ALGORITHM_CHANGE = true;
     private static final Boolean DEFAULT_MIGRATE_WITH_BACKUP = false;
-    private static final Boolean DEFAULT_ENCRYPTED_SHARED_PREFERENCES = false;
     private static final Boolean DEFAULT_ENFORCE_BIOMETRICS = false;
     private static final String DEFAULT_BIOMETRIC_TYPE = BIOMETRIC_TYPE_DEVICE_CREDENTIAL;
     private static final String DEFAULT_BIOMETRIC_PROMPT_TITLE = "Authenticate to access";
@@ -31,7 +30,6 @@ public class FlutterSecureStorageConfig {
     public static final String PREF_OPTION_DELETE_ON_FAILURE = "resetOnError";
     public static final String PREF_OPTION_MIGRATE_ON_ALGORITHM_CHANGE = "migrateOnAlgorithmChange";
     public static final String PREF_OPTION_MIGRATE_WITH_BACKUP = "migrateWithBackup";
-    public static final String PREF_OPTION_ENCRYPTED_SHARED_PREFERENCES = "encryptedSharedPreferences";
     public static final String PREF_OPTION_ENFORCE_BIOMETRICS = "enforceBiometrics";
     public static final String PREF_OPTION_BIOMETRIC_TYPE = "biometricType";
     public static final String PREF_OPTION_BIOMETRIC_PROMPT_TITLE = "biometricPromptTitle";
@@ -53,7 +51,6 @@ public class FlutterSecureStorageConfig {
     private final boolean deleteOnFailure;
     private final boolean migrateOnAlgorithmChange;
     private final boolean migrateWithBackup;
-    private final boolean useEncryptedSharedPreferences;
     private final boolean enforceBiometrics;
     private final boolean strongBiometricOnly;
     private final String biometricPromptTitle;
@@ -68,7 +65,6 @@ public class FlutterSecureStorageConfig {
         this.deleteOnFailure = getBooleanOption(options, PREF_OPTION_DELETE_ON_FAILURE, DEFAULT_DELETE_ON_FAILURE);
         this.migrateOnAlgorithmChange = getBooleanOption(options, PREF_OPTION_MIGRATE_ON_ALGORITHM_CHANGE, DEFAULT_MIGRATE_ON_ALGORITHM_CHANGE);
         this.migrateWithBackup = getBooleanOption(options, PREF_OPTION_MIGRATE_WITH_BACKUP, DEFAULT_MIGRATE_WITH_BACKUP);
-        this.useEncryptedSharedPreferences = getBooleanOption(options, PREF_OPTION_ENCRYPTED_SHARED_PREFERENCES, DEFAULT_ENCRYPTED_SHARED_PREFERENCES);
         this.enforceBiometrics = getBooleanOption(options, PREF_OPTION_ENFORCE_BIOMETRICS, DEFAULT_ENFORCE_BIOMETRICS);
         String biometricTypeValue = getStringOption(options, PREF_OPTION_BIOMETRIC_TYPE, DEFAULT_BIOMETRIC_TYPE);
         if (!BIOMETRIC_TYPE_STRONG.equals(biometricTypeValue) && !BIOMETRIC_TYPE_DEVICE_CREDENTIAL.equals(biometricTypeValue)) {
@@ -150,7 +146,6 @@ public class FlutterSecureStorageConfig {
     public boolean shouldMigrateOnAlgorithmChange() { return migrateOnAlgorithmChange; }
     public boolean shouldMigrateWithBackup() { return migrateWithBackup; }
 
-    public boolean isUseEncryptedSharedPreferences() { return useEncryptedSharedPreferences; }
     public boolean getEnforceBiometrics() { return enforceBiometrics; }
     public boolean isStrongBiometricOnly() { return strongBiometricOnly; }
