+++ flutter_secure_storage/android/src/main/java/com/it_nomads/fluttersecurestorage/ciphers/KeyCipherAlgorithm.java
@@ -3,7 +3,6 @@ package com.it_nomads.fluttersecurestorage.ciphers;
 import android.os.Build;
 
 public enum KeyCipherAlgorithm {
-    RSA_ECB_PKCS1Padding(KeyCipherImplementationRSA18::new, 1),
     RSA_ECB_OAEPwithSHA_256andMGF1Padding(KeyCipherImplementationRSAOAEP::new, Build.VERSION_CODES.M),
     AES_GCM_NoPadding(KeyCipherImplementationAES23::new, Build.VERSION_CODES.M); // Renamed from AES_GCM_NoPadding_BIOMETRIC
     final KeyCipherFunction keyCipher;
