+++ flutter_secure_storage/android/src/main/java/com/it_nomads/fluttersecurestorage/ciphers/StorageCipherFactory.java
@@ -13,8 +13,8 @@ public class StorageCipherFactory {
     private static final String ELEMENT_PREFERENCES_ALGORITHM_PREFIX = "FlutterSecureSAlgorithm";
     private static final String ELEMENT_PREFERENCES_ALGORITHM_KEY = ELEMENT_PREFERENCES_ALGORITHM_PREFIX + "Key";
     private static final String ELEMENT_PREFERENCES_ALGORITHM_STORAGE = ELEMENT_PREFERENCES_ALGORITHM_PREFIX + "Storage";
-    private static final KeyCipherAlgorithm DEFAULT_KEY_ALGORITHM = KeyCipherAlgorithm.RSA_ECB_PKCS1Padding;
-    private static final StorageCipherAlgorithm DEFAULT_STORAGE_ALGORITHM = StorageCipherAlgorithm.AES_CBC_PKCS7Padding;
+    private static final KeyCipherAlgorithm DEFAULT_KEY_ALGORITHM = KeyCipherAlgorithm.RSA_ECB_OAEPwithSHA_256andMGF1Padding;
+    private static final StorageCipherAlgorithm DEFAULT_STORAGE_ALGORITHM = StorageCipherAlgorithm.AES_GCM_NoPadding;
 
     private final KeyCipherAlgorithm savedKeyAlgorithm;
     private final StorageCipherAlgorithm savedStorageAlgorithm;
@@ -28,14 +28,11 @@ public class StorageCipherFactory {
         final String savedStorageCipherAlgorithm = configSource.getString(ELEMENT_PREFERENCES_ALGORITHM_STORAGE, null);
 
         if (savedKeyCipherAlgorithm == null || savedStorageCipherAlgorithm == null) {
-            // Migration from v9.2.4 or v10.0.0-beta.4:
-            // No algorithm markers exist in SharedPreferences, which means the data was encrypted
-            // with the historical v9.2.4 defaults. We must use these defaults to decrypt the old
-            // data, even if the current config specifies different algorithms.
-            // After successful decryption, the data will be re-encrypted with current algorithms
-            // (if they differ) via the migration flow in handleKeyMismatch().
-            savedKeyAlgorithm = DEFAULT_KEY_ALGORITHM;        // RSA_ECB_PKCS1Padding
-            savedStorageAlgorithm = DEFAULT_STORAGE_ALGORITHM; // AES_CBC_PKCS7Padding
+            // No algorithm markers exist, treat as a fresh install using current defaults.
+            // v11 requires users to have migrated through v10 first; data from v9 or earlier
+            // without markers is treated as if it uses the current defaults.
+            savedKeyAlgorithm = DEFAULT_KEY_ALGORITHM;
+            savedStorageAlgorithm = DEFAULT_STORAGE_ALGORITHM;
         } else {
             savedKeyAlgorithm = KeyCipherAlgorithm.fromString(savedKeyCipherAlgorithm);
             savedStorageAlgorithm = StorageCipherAlgorithm.fromString(savedStorageCipherAlgorithm);
