+++ flutter_secure_storage/android/src/test/java/com/it_nomads/fluttersecurestorage/FlutterSecureStorageConfigTest.java
@@ -51,11 +51,6 @@ public class FlutterSecureStorageConfigTest {
     }
 
     @Test
-    public void defaults_encryptedSharedPreferences_isFalse() {
-        assertFalse(emptyConfig().isUseEncryptedSharedPreferences());
-    }
-
-    @Test
     public void defaults_enforceBiometrics_isFalse() {
         assertFalse(emptyConfig().getEnforceBiometrics());
     }
@@ -117,12 +112,6 @@ public class FlutterSecureStorageConfigTest {
     }
 
     @Test
-    public void custom_encryptedSharedPreferences_true() {
-        FlutterSecureStorageConfig config = configFrom(FlutterSecureStorageConfig.PREF_OPTION_ENCRYPTED_SHARED_PREFERENCES, "true");
-        assertTrue(config.isUseEncryptedSharedPreferences());
-    }
-
-    @Test
     public void custom_enforceBiometrics_true() {
         FlutterSecureStorageConfig config = configFrom(FlutterSecureStorageConfig.PREF_OPTION_ENFORCE_BIOMETRICS, "true");
         assertTrue(config.getEnforceBiometrics());
@@ -177,8 +166,8 @@ public class FlutterSecureStorageConfigTest {
 
     @Test
     public void custom_storageCipherAlgorithm() {
-        FlutterSecureStorageConfig config = configFrom(FlutterSecureStorageConfig.PREF_OPTION_STORAGE_CIPHER_ALGORITHM, "AES_CBC_PKCS7Padding");
-        assertEquals("AES_CBC_PKCS7Padding", config.getPrefOptionStorageCipherAlgorithm());
+        FlutterSecureStorageConfig config = configFrom(FlutterSecureStorageConfig.PREF_OPTION_STORAGE_CIPHER_ALGORITHM, "AES_GCM_NoPadding");
+        assertEquals("AES_GCM_NoPadding", config.getPrefOptionStorageCipherAlgorithm());
     }
 
     @Test
