+++ flutter_secure_storage/android/src/test/java/com/it_nomads/fluttersecurestorage/ciphers/CipherAlgorithmTest.java
@@ -12,11 +12,6 @@ public class CipherAlgorithmTest {
     // -------------------------------------------------------------------------
 
     @Test
-    public void keyCipher_fromString_RSA_ECB_PKCS1Padding() {
-        assertEquals(KeyCipherAlgorithm.RSA_ECB_PKCS1Padding, KeyCipherAlgorithm.fromString("RSA_ECB_PKCS1Padding"));
-    }
-
-    @Test
     public void keyCipher_fromString_RSA_ECB_OAEPwithSHA_256andMGF1Padding() {
         assertEquals(
             KeyCipherAlgorithm.RSA_ECB_OAEPwithSHA_256andMGF1Padding,
@@ -48,14 +43,15 @@ public class CipherAlgorithmTest {
         assertThrows(IllegalArgumentException.class, () -> KeyCipherAlgorithm.fromString(""));
     }
 
+    @Test
+    public void keyCipher_fromString_removedPKCS1_throwsIllegalArgumentException() {
+        // RSA_ECB_PKCS1Padding was removed in v11
+        assertThrows(IllegalArgumentException.class, () -> KeyCipherAlgorithm.fromString("RSA_ECB_PKCS1Padding"));
+    }
+
     // -------------------------------------------------------------------------
     // StorageCipherAlgorithm.fromString
     // -------------------------------------------------------------------------
-
-    @Test
-    public void storageCipher_fromString_AES_CBC_PKCS7Padding() {
-        assertEquals(StorageCipherAlgorithm.AES_CBC_PKCS7Padding, StorageCipherAlgorithm.fromString("AES_CBC_PKCS7Padding"));
-    }
 
     @Test
     public void storageCipher_fromString_AES_GCM_NoPadding() {
@@ -81,17 +77,23 @@ public class CipherAlgorithmTest {
         assertThrows(IllegalArgumentException.class, () -> StorageCipherAlgorithm.fromString(""));
     }
 
+    @Test
+    public void storageCipher_fromString_removedCBC_throwsIllegalArgumentException() {
+        // AES_CBC_PKCS7Padding was removed in v11
+        assertThrows(IllegalArgumentException.class, () -> StorageCipherAlgorithm.fromString("AES_CBC_PKCS7Padding"));
+    }
+
     // -------------------------------------------------------------------------
     // Enum completeness — guards against accidental removal of values
     // -------------------------------------------------------------------------
 
     @Test
     public void keyCipher_hasExpectedNumberOfValues() {
-        assertEquals(3, KeyCipherAlgorithm.values().length);
+        assertEquals(2, KeyCipherAlgorithm.values().length);
     }
 
     @Test
     public void storageCipher_hasExpectedNumberOfValues() {
-        assertEquals(2, StorageCipherAlgorithm.values().length);
+        assertEquals(1, StorageCipherAlgorithm.values().length);
     }
 }
