+++ flutter_secure_storage/lib/options/android_options.dart
@@ -5,16 +5,6 @@ part of '../flutter_secure_storage.dart';
 
 /// Algorithm used to encrypt/wrap the secret key in Android KeyStore.
 enum KeyCipherAlgorithm {
-  /// Legacy RSA/ECB/PKCS1Padding for backwards compatibility.
-  ///
-  /// PKCS#1 v1.5 padding is vulnerable to padding-oracle attacks.
-  /// Existing data will be automatically migrated to the default algorithm
-  /// on first access when `migrateOnAlgorithmChange` is true (the default).
-  @Deprecated('RSA PKCS#1 v1.5 padding is insecure. '
-      'Use the default RSA_ECB_OAEPwithSHA_256andMGF1Padding instead. '
-      'Existing data is migrated automatically.')
-  RSA_ECB_PKCS1Padding,
-
   /// RSA/ECB/OAEPWithSHA-256AndMGF1Padding (default, API 23+).
   RSA_ECB_OAEPwithSHA_256andMGF1Padding,
 
@@ -24,18 +14,6 @@ enum KeyCipherAlgorithm {
 
 /// Algorithm used to encrypt stored data.
 enum StorageCipherAlgorithm {
-  /// Legacy AES/CBC/PKCS7Padding for backwards compatibility.
-  ///
-  /// CBC mode is vulnerable to padding-oracle attacks and does not provide
-  /// authentication. Existing data will be automatically migrated to the
-  /// default algorithm on first access when `migrateOnAlgorithmChange` is
-  /// true (the default).
-  @Deprecated(
-      'AES-CBC is insecure (no authentication, padding-oracle vulnerable). '
-      'Use the default AES_GCM_NoPadding instead. '
-      'Existing data is migrated automatically.')
-  AES_CBC_PKCS7Padding,
-
   /// AES/GCM/NoPadding (default, API 23+).
   AES_GCM_NoPadding,
 }
@@ -65,18 +43,11 @@ class AndroidOptions extends Options {
   ///
   /// For biometric authentication, use `AndroidOptions.biometric()`.
   ///
-  /// Advanced users can customize cipher algorithms for specific use cases.
-  /// Valid combinations:
-  /// - AES_CBC_PKCS7Padding storage + any key cipher
+  /// Advanced users can customize cipher algorithms for specific use cases:
   /// - AES_GCM_NoPadding storage + RSA key ciphers (standard RSA wrapping)
   /// - AES_GCM_NoPadding storage + AES_GCM_NoPadding key
   ///   (KeyStore-based, supports biometrics)
   const AndroidOptions({
-    @Deprecated('EncryptedSharedPreferences is deprecated and will be '
-        'removed in v11. The Jetpack Security library is deprecated by Google. '
-        'Your data will be automatically migrated to custom ciphers on first '
-        'access. Remove this parameter - it will be ignored.')
-    bool encryptedSharedPreferences = false,
     bool resetOnError = true,
     bool migrateOnAlgorithmChange = true,
     bool migrateWithBackup = false,
@@ -87,18 +58,12 @@ class AndroidOptions extends Options {
         StorageCipherAlgorithm.AES_GCM_NoPadding,
     AndroidBiometricType biometricType =
         AndroidBiometricType.biometricOrDeviceCredential,
-    @Deprecated(
-        'Use storageNamespace instead. sharedPreferencesName only isolates '
-        'data storage; storageNamespace provides full isolation including '
-        'KeyStore aliases and key storage.')
-    this.sharedPreferencesName,
     this.preferencesKeyPrefix,
     this.storageNamespace,
     this.biometricPromptTitle,
     this.biometricPromptSubtitle,
     this.biometricPromptNegativeButton,
-  })  : _encryptedSharedPreferences = encryptedSharedPreferences,
-        _resetOnError = resetOnError,
+  })  : _resetOnError = resetOnError,
         _migrateOnAlgorithmChange = migrateOnAlgorithmChange,
         _migrateWithBackup = migrateWithBackup,
         _enforceBiometrics = enforceBiometrics,
@@ -115,38 +80,24 @@ class AndroidOptions extends Options {
   /// - When enforceBiometrics=false, gracefully degrades if biometrics
   ///   unavailable
   const AndroidOptions.biometric({
-    @Deprecated(
-        'EncryptedSharedPreferences is deprecated and will be removed in v11. '
-        'The Jetpack Security library is deprecated by Google. '
-        'Remove this parameter - it will be ignored.')
-    bool encryptedSharedPreferences = false,
     bool resetOnError = true,
     bool migrateOnAlgorithmChange = true,
     bool migrateWithBackup = false,
     bool enforceBiometrics = false,
     AndroidBiometricType biometricType =
         AndroidBiometricType.biometricOrDeviceCredential,
-    @Deprecated(
-        'Use storageNamespace instead. sharedPreferencesName only isolates '
-        'data storage; storageNamespace provides full isolation including '
-        'KeyStore aliases and key storage.')
-    this.sharedPreferencesName,
     this.preferencesKeyPrefix,
     this.storageNamespace,
     this.biometricPromptTitle,
     this.biometricPromptSubtitle,
     this.biometricPromptNegativeButton,
-  })  : _encryptedSharedPreferences = encryptedSharedPreferences,
-        _resetOnError = resetOnError,
+  })  : _resetOnError = resetOnError,
         _migrateOnAlgorithmChange = migrateOnAlgorithmChange,
         _migrateWithBackup = migrateWithBackup,
         _enforceBiometrics = enforceBiometrics,
         _keyCipherAlgorithm = KeyCipherAlgorithm.AES_GCM_NoPadding,
         _storageCipherAlgorithm = StorageCipherAlgorithm.AES_GCM_NoPadding,
         _biometricType = biometricType;
-
-  /// EncryptedSharedPrefences are only available on API 23 and greater
-  final bool _encryptedSharedPreferences;
 
   /// When an error is detected, automatically reset all data. This will prevent
   /// fatal errors regarding an unknown key however keep in mind that it will
@@ -189,28 +140,16 @@ class AndroidOptions extends Options {
 
   /// Algorithm used to encrypt the secret key.
   /// By default RSA/ECB/OAEPWithSHA-256AndMGF1Padding is used (API 23+).
-  /// Legacy RSA/ECB/PKCS1Padding is available for backwards compatibility.
   final KeyCipherAlgorithm _keyCipherAlgorithm;
 
   /// Algorithm used to encrypt stored data.
   /// By default AES/GCM/NoPadding is used (API 23+).
-  /// Legacy AES/CBC/PKCS7Padding is available for backwards compatibility.
   final StorageCipherAlgorithm _storageCipherAlgorithm;
 
   /// Controls which authentication methods are accepted during biometric
   /// prompts.
   final AndroidBiometricType _biometricType;
 
-  /// The name of the sharedPreference database to use.
-  /// You can select your own name if you want. A default name will
-  /// be used if nothing is provided here.
-  ///
-  /// WARNING: If you change this you can't retrieve already saved preferences.
-  @Deprecated(
-      'Use storageNamespace instead. sharedPreferencesName only isolates '
-      'data storage; storageNamespace provides full isolation including '
-      'KeyStore aliases and key storage.')
-  final String? sharedPreferencesName;
 
   /// The prefix for a shared preference key. The prefix is used to make sure
   /// the key is unique to your application. An underscore (_) is added to the
@@ -234,7 +173,8 @@ class AndroidOptions extends Options {
   /// This allows multiple `FlutterSecureStorage` instances to use different
   /// cipher algorithms without conflicting KeyStore entries or key storage.
   ///
-  /// Prefer this over `sharedPreferencesName` for new code.
+  /// Prefer this over the legacy `sharedPreferencesName` (removed in v11)
+  /// for namespace isolation.
   final String? storageNamespace;
 
   /// The title shown in the biometric authentication prompt.
@@ -255,7 +195,6 @@ class AndroidOptions extends Options {
 
   @override
   Map<String, String> toMap() => <String, String>{
-        'encryptedSharedPreferences': '$_encryptedSharedPreferences',
         'resetOnError': '$_resetOnError',
         'migrateOnAlgorithmChange': '$_migrateOnAlgorithmChange',
         'migrateWithBackup': '$_migrateWithBackup',
@@ -263,8 +202,6 @@ class AndroidOptions extends Options {
         'keyCipherAlgorithm': _keyCipherAlgorithm.name,
         'storageCipherAlgorithm': _storageCipherAlgorithm.name,
         'biometricType': _biometricType.name,
-        // ignore: deprecated_member_use_from_same_package — legacy support
-        'sharedPreferencesName': sharedPreferencesName ?? '',
         'preferencesKeyPrefix': preferencesKeyPrefix ?? '',
         'storageNamespace': storageNamespace ?? '',
         'biometricPromptTitle':
@@ -277,7 +214,6 @@ class AndroidOptions extends Options {
 
   /// Creates a copy of this AndroidOptions with the given fields replaced.
   AndroidOptions copyWith({
-    bool? encryptedSharedPreferences,
     bool? resetOnError,
     bool? migrateOnAlgorithmChange,
     bool? migrateWithBackup,
@@ -286,20 +222,12 @@ class AndroidOptions extends Options {
     StorageCipherAlgorithm? storageCipherAlgorithm,
     AndroidBiometricType? biometricType,
     String? preferencesKeyPrefix,
-    @Deprecated(
-        'Use storageNamespace instead. sharedPreferencesName only isolates '
-        'data storage; storageNamespace provides full isolation including '
-        'KeyStore aliases and key storage.')
-    String? sharedPreferencesName,
     String? storageNamespace,
     String? biometricPromptTitle,
     String? biometricPromptSubtitle,
     String? biometricPromptNegativeButton,
   }) =>
       AndroidOptions(
-        // ignore: deprecated_member_use_from_same_package — will be removed in v11
-        encryptedSharedPreferences:
-            encryptedSharedPreferences ?? _encryptedSharedPreferences,
         resetOnError: resetOnError ?? _resetOnError,
         migrateOnAlgorithmChange:
             migrateOnAlgorithmChange ?? _migrateOnAlgorithmChange,
@@ -309,10 +237,6 @@ class AndroidOptions extends Options {
         storageCipherAlgorithm:
             storageCipherAlgorithm ?? _storageCipherAlgorithm,
         biometricType: biometricType ?? _biometricType,
-        // ignore: deprecated_member_use_from_same_package — legacy support
-        sharedPreferencesName:
-            // ignore: deprecated_member_use_from_same_package — legacy support
-            sharedPreferencesName ?? this.sharedPreferencesName,
         preferencesKeyPrefix: preferencesKeyPrefix ?? this.preferencesKeyPrefix,
         storageNamespace: storageNamespace ?? this.storageNamespace,
         biometricPromptTitle: biometricPromptTitle ?? this.biometricPromptTitle,
